+++ src/app/app.module.ts
@@ -2,26 +2,36 @@ import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 
 import {AppComponent} from './app.component';
-import {ProductComponentComponent} from './product-component/product-component.component';
-import {ProductListComponentComponent} from './product-list-component/product-list-component.component';
+import {ProductComponentComponent} from './components/product-component/product-component.component';
+import {ProductListComponentComponent} from './components/product-list-component/product-list-component.component';
 import {ProductService} from "./services/product/product.service";
-import {CartComponentComponent} from './cart-component/cart-component.component';
-import {CartItemComponent} from './cart-component/cart-item/cart-item.component';
-import {SharedModule} from "./shared/shared.module";
+import {CartComponentComponent} from './components/cart-component/cart-component.component';
+import {CartItemComponent} from './components/cart-component/cart-item/cart-item.component';
+import {CoreModule} from "./shared/core.module";
+import { ContactUsComponent } from './components/contact-us/contact-us.component';
+import {ConfigOptionsService} from "./shared/services/config-options.service";
+import {LocalStorageService} from "./shared/services/local-storage.service";
+import {GeneratorService} from "./shared/services/generator.service";
+import {Constant, ConstantsService} from "./shared/services/constantsService";
 
 @NgModule({
   imports: [
     BrowserModule,
-    SharedModule
+    CoreModule
   ],
   declarations: [
     AppComponent,
     ProductComponentComponent,
     ProductListComponentComponent,
     CartComponentComponent,
-    CartItemComponent
+    CartItemComponent,
+    ContactUsComponent
   ],
-  providers: [ProductService],
+  providers: [ProductService, LocalStorageService,
+    ConfigOptionsService,
+    GeneratorService,
+    {provide: ConstantsService, useValue: Constant}
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {