+++ src/app/shared/services/index.ts
@@ -1,0 +1,5 @@
+export {ConstantsService} from "./constantsService";
+export {Constant} from "./constantsService";
+export {ConfigOptionsService} from "./config-options.service";
+export {LocalStorageService} from "./local-storage.service";
+export {GeneratorService} from "./generator.service";